deprecated/aggTrades/historical/transformation/src/requirements.txt
boto3
pandas==2.0.0
pyspark==3.5.1
python-dotenv==1.0.1
pytz==2023.3.post1
Requests==2.32.3
schedule==1.2.2
